Release the project under the standard Rust permissive licenses
Dual-licensing the project under either the MIT or Apache 2.0 licenses
provides for maximum compatibility within the Rust ecosystem and is
recommended by the upstream API guidelines.

Copyrights in the project are retained by their contributors and no
copyright assignment is required to contribute. Authorship information
can be found in the version control history.

See:
- https://rust-lang.github.io/api-guidelines/necessities.html#crate-and-its-dependencies-have-a-permissive-license-c-permissive