Rename "CI workflow" to "test workflow"
I think this is a bit more clear, and will allow for other logical
workflows like "check", "release", etc. Trying it on for size. When
there a lot of commit statuses posted, it's a bit easier to read if
everything is lowercase, although run steps are still capitalized to
match the builtin step formatting.

See:
- https://youtu.be/xUH-4y92jPg