Always run the workflow status summary job
This is used by bors to help it determine the overall status once job
structure gains some complexity. I've also found it confusing to see the
job skipped on PR tests. I know some of the rust-lang project define two
separate jobs with the same `name:`, but I that's a bit ugly and still
has one that is always skipped. I think this will ensure that it reports
success/failure as we want it to. I don't think we need to check that it
was bors that did the pushing, so that's been removed.