Ensure the desired Rust toolchain is installed
I know this is a third-party action that is not pinned to a revision,
but it's relatively simple and works by parsing that out directly. I
could, in theory pin it and specify the toolchain as an input, but the
upstream repo doesn't add tags/releases, so dependabot won't help to
bump versions.

I'm using this over the common `action-rs/toolchain`, as it is
unmaintained.

See:
- https://github.com/dtolnay/rust-toolchain
- https://github.com/actions-rs/toolchain/issues/216
- https://old.reddit.com/r/rust/comments/vyx4oj/actionsrs_organization_became_unmaintained/ig54zv7/