Run tests on the main branch to establish caches
While this may be technically redundant since staging is what
essentially becomes main after a successful build, caches can be shared
on PRs from the target branch (main), so we'll want to ensure that we
build there. It also makes sense that we'll want our CI status badge to
point to these builds to prevent confusion.