Clean up README formatting a bit
Since we now have multiple GitHub Actions workflows, we need more badge
sophistication than what GitHub provides out of the box. We want it to
show the status of checks that have run on the main branch and link to
all actions rather than a specific workflow. I took most of the style
from Serde's repo.

I have a visceral aversion to the copy/paste format of the licensing
section from the API docs. This version is much cleaner to me, very
similar to what I've put in my personal repos for years, and basically
what many of the rust-lang projects use themselves.

See:
- https://github.com/serde-rs/serde
- https://rust-lang.github.io/api-guidelines/necessities.html#crate-and-its-dependencies-have-a-permissive-license-c-permissive
- https://github.com/rust-lang/rustfmt#license