First cut of imported CUE workflows and genworkflows cmd
First we generated CUE base on the existing YAML files:

    $ cd workflows/
    $ cue import -p workflows -l 'name' *.yml

After writing a list of all the workflows and the desired filenames, we
use CUE to generate the YAML files. The only differences to the original
were the stripping of comments (I'll add them back to the CUE files
soon), and some minor YAML syntax changes for consistency.

    $ cue cmd genworkflows