Add constraint on github.#Workflow
This ensures that our workflow output matches the upstream JSONSchema. I
also copied the CUE project's vendoring code to automatically import the
workflow schema so it won't have to be a manual process.

Note that this is not the most recent commit of the schema, as the
commit after this one introduced a change that CUE doesn't handle
properly, and running `cue vet -c` will spit out an "incomplete value"
error for every job step.

Instead, I'm matching what the CUE project is doing here and hope
they will fix their parsing soon. I've searched for a known issue, but
haven't found one yet. I'll keep looking and will report it for tracking
if I can't track it down.

See:
- https://github.com/cue-lang/cue/blob/59080b65a5ba539bb3f3cb45d3346f39df67cf01/internal/ci/vendor/vendor_tool.cue#L43
- https://github.com/SchemaStore/schemastore/commits/b83afed7ff89fb17717394369eb1ba9684bb20d2/src/schemas/json/github-workflow.json