Pin the actions/checkout version
I was previously thinking that it's okay to trust the GitHub-managed
actions, but it's not just about trust, it's also about being able
to run an older version of jobs if checking out a previous revision.
Dependabot will also give us notices when upstream action providers
tag new releases, so it shouldn't be too hard to maintain. Additional
tooling could also be written to help, if needed.