Only run Rust workflow on pushes to main branch
The other workflows are mostly check/format/lint items for the rest of
the repo, so they should be just fine. I know the rust-cache option can
pull from the target branch (main) when running on PRs, but I'm not sure
the other tooling benefits from it being present in the main branch.