Add initial xtask structure for use as Makefile replacement
This is a conventional structure used to replace Bash scripts and Make
with pure Rust to handle common glue tasks. I want developers to be able
to easily run commands that invoke the various non-Rust tools that are
used to keep the repository in good shape.

Eventually, there will be a single command to run (`cargo xtask ci`),
where the user can expect that, if it passes locally, the CI will pass
as well.

See:
- https://github.com/matklad/cargo-xtask