Use xshell for xtask command line scripting
This provides a much nicer interface for handling environmental changes
and calling commands in a safe manner. It will also provide friendly
messages if a command is missing (so we won't need explicit dependency
checks), and it will automatically output the command that it's running,
for reference. It's a simple, but well-designed library for developing
our xtasks.

See:
- https://docs.rs/xshell/latest/xshell/