No need to run the rust workflow on the main branch
With the merge queue setup, it has already ran and passed on the PR
itself, and then a second time in the merge queue, where the commit
should retain it's checks status. I'll double-check that with our
current caching setup, we don't have an issue, but I think this should
still work as expected.