Keep Rust caches on main branch warm
Due to cache isolation for security, workflow runs can only restore
caches created in either the current branch, the base branch for a PR,
or the repo's default branch. Since we're now using merge queues, we
don't have regular jobs running on pushes to the main branch, so we
won't have a warm cache that is accessible by pull requests. Running
check's on each of the supported Rust versions and operating systems
will be enough to preload the caches and keep our jobs running more
efficiently.

GitHub will remove any cache entries that have not been accessed in over
7 days, and we'd also like to ensure that we always cache the latest
Rust stable version even if our code/dependencies haven't changed, so
we'll also preload the cache on a weekly basis. High load times for
GitHub Actions include the start of every hour, so we'll offset that
a bit.

See:
- https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#restrictions-for-accessing-a-cache
- https://github.com/Swatinem/rust-cache#cache-details
- https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy
- https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule