Make sure that cargo-nextest is cached as well
The order matters here since rust-cache cleans out any files in
`~/.cargo/bin` that were present **before the action ran**, which means
we'll want to install nextest **after** running the cache job. The
install-action job will check for newer versions appropriately.

The one thing I'm not sure about is if rust-cache will key based off
of the contents of the `~/.cargo/bin` directory, or if it's just saving
that into the cache. If it does key off of the contents, we may need
to always install nextest even if it's not going to be used (for check/
format/lint jobs) so we can maintain cache efficiency.

See:
- https://github.com/Swatinem/rust-cache#cache-details
- https://github.com/taiki-e/install-action/issues/66