Allow clippy runs during the fixup.rust xtask to error out
When running clippy with the `--fix` flag, it will demote all errors
into warnings rather than halt if it can't fix the issue automatically.
To make sure that we actually address any clippy lints, we need to
re-run it afterward without the `--fix` flag, but ensuring that all
warnings are treated as errors.

Note that I've removed the `--no-deps` flag, as for clippy, that is used
when you only want to test a single package within a workspace, not for
looking at third-party dependencies.

We've also added in a `cargo fix` step to handle changes that the
compiler can do itself (outside of clippy), while preferring to adopt
any of the current Rust edition's idioms.

See:
- https://github.com/rust-lang/rust-clippy/discussions/9979
- https://doc.rust-lang.org/stable/clippy/usage.html#workspaces
- https://doc.rust-lang.org/cargo/commands/cargo-fix.html