Timeout the Rust cache step after five minutes
I had an instance where, for whatever reason, it got stuck restoring
the cache and took an hour before it timed out and moved on. I tried
to cancel the job run, but it did not have an effect, so everything was
stuck until the timeout was reached. It should be fine if the cache is
missed, as it's purely an optimization, but I want to wait long enough
where I'd probably notice if it starts timing out regularly.

This was the output on the job:

    [...]
    Received 167485073 of 171679377 (97.6%), 1.6 MBs/sec
    Received 167485073 of 171679377 (97.6%), 1.6 MBs/sec
    Warning: Failed to restore: The operation cannot be completed in timeout.
    No cache found.