Use lexopt for minimalist argument parsing
I compared a lot of options for minimalist argument parsing here
and really liked the approach that lexopt takes. It's super easy to
understand and provides enough functionality to where I can expand
things if necessary for future xtasks. The philosophy of the developer
is also in-line with what I like.

I'm also now looping over the args, so more than one command can be given.

See:
- https://github.com/blyxxyz/lexopt
- https://github.com/rosetta-rs/argparse-rosetta-rs