Don't shell out for removing files
Using Rust's `std::fs::remove_file` function should be safer and
platform-independent. Since I am not going to display the paths to the
files we're removing and want to eliminate the shell's current directory
from the equation, I split out the relative_paths functionality into
a dedicated function, so we can get either absolute or relative as
necessary.