Show module structure integration tests
This is a trivial example of how to structure integration tests for
mylib. It's not super realistic, but it's just there for demonstration.

I'm holding off on moving the unit tests module into a dedicated
file...it is common to see on larger crates, and cargo can avoid
unnecessary compilation depending on which file was changed, but it's
a bit of premature optimization? I'll have to think about it some more.

See:
- https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html