Add separate step to compile Miri tests
This makes timing easier to compare across runs since you can account
for compilation versus the test runs themselves. Note that I'm not going
to add `cargo-nextest` into the mix because it prevents Miri from being
able to detect data races where two tests race on a shared resource.
This is a scheduled job anyway, so speed is less of a concern.

See:
- https://github.com/rust-lang/miri#running-tests-in-parallel