Add an MSRV policy to the README
This is a bit of a tricky topic, as the Rust ecosystem is waiting on
better Cargo support to make usability better, and there are a lot of
differing opinions and theories on how best to handle an MSRV. Not a
lot of concrete data out there. The Rust compiler itself explicitly only
supports the latest stable version, so any MSRV policy is essentially
how long you're willing to wait for dependents of your library to
upgrade before you adopt new features. Artificially holding this back
might also mean missing out on security fixes, and I'm a big fan of
staying current when possible.

This can always be changed, but I wanted to have something in place to
make sure that it's covered.

See:
- https://github.com/rust-lang/libs-team/issues/72
- https://github.com/time-rs/time/discussions/535