Add --ignore-missing option flag
This allows a user to skip any tasks that call commands that aren't
installed on their system. If a command is missing, special care is
taken to display a link to help the user install the missing tool. There
is a safety check to never allow skipping when xtasks are run in a CI
environment.