Add an initial dist xtask for packaging release artifacts
There's more work to be done here, but it's a start. Honestly, adding
miniserde just to parse out the Cargo metadata seems way more difficult
than it should be to get an accurate list of the binary artifacts for
the project. There is an unstable `--out-dir` feature to copy artifacts
for you after building, but it hasn't stabilized over the past couple of
years. Not ideal.

See:
- https://github.com/rust-lang/cargo/issues/3757
- https://github.com/rust-lang/cargo/issues/6790#issuecomment-1488959163
- https://github.com/matklad/cargo-xtask/issues/3