Fix non-binding `let` without a type annotation
This is a pedantic lint in the current stable Rust, but now a restricted
lint in nightly due to some known issues. In our case, I had just taken
it from an xtask examples file without thinking about the return just
being a Result<()>, which we're already handling with anyhow.

See:
- https://rust-lang.github.io/rust-clippy/master/index.html#let_underscore_untyped
- https://doc.rust-lang.org/std/fs/fn.remove_dir_all.html