Rely on deny warnings to catch all configured clippy lints
Now that we put the various lint categories into the files themselves,
we can just rely on CI to deny all warnings rather than explicitly
adding the pedantic category. This should make it more maintainable
since if we need exceptions, we can just put them in the source files
directly and not cause issues.