Use nightly rustfmt with unstable config options
These are really nice config settings, and although they're mostly not
stable, we can adjust our process to use nightly for running `cargo
fmt`.

Note that I'm getting rid of the code to honor the $CARGO environment
variable. It's not a common practice to override it, and in our case,
it made switching between toolchains with the override shorthand syntax
awkward because running `cargo xtask ...` would set the environment
variable to point to the stable toolchain's absolute path. Trying to
add `+nightly` to that doesn't work, as there's normally a wrapper shim
that decides the version to execute beforehand. It also made the output
more confusing to see the whole path in there, and doubly so trying to
override things using the $RUSTUP_TOOLCHAIN environment variable instead
of the shorthand syntax.