Add production profile for maximum optimization
This is what we'll use to generate the distributable artifacts. Some of
these things don't exist on the default release profile because there's
an aversion to making compilation times longer.

This also moves `strip` to just the production profile and adds "thin"
LTO to the release profile. It'll be nice to have a faster version of
LTO for local release builds (and benchmarks), but it won't add too much
time for compiling.

See:
- https://doc.rust-lang.org/cargo/reference/profiles.html
- https://nnethercote.github.io/perf-book/build-configuration.html
- https://github.com/rust-lang/cargo/issues/11298