Switch from codespell to typos for spell checking
These are very similar tools, but typos is written in Rust (versus
Python), and should be easier for users to install. It's also quite a
bit faster and will hopefully require less tweaking. When testing it
on some other repos, it caught some typos that codespell did not, but
that's mostly due to differences in project dictionaries. That said,
it does do some filetype detection and accounts for language-specific
quirks, as well as the distinction between identifiers vs words, which
is great.

See:
- https://github.com/crate-ci/typos