Add build script to inject Git built-time metadata
This allows our application to pull useful information from the
environment when compiling, so we can have build metadata info displayed
for version information, etc..

The format is based on what rustc does; it will display the commit
hash and commit date, but additionally will add a `.dirty` flag if the
repository was not clean.

Using 12 characters to abbreviate the Git hash should be more than
sufficient to be unambiguous.

See:
- https://blog.cuviper.com/2013/11/10/how-short-can-git-abbreviate/