Restructure xtask dist code
I missed that the switch to the production profile actually broke
compilation, because we didn't update all references accordingly. I've
also tweaked how the xtask crate itself is ignored, and added in the
ability to copy project docs into the dist directory.

For now, I'm leaving the generate_assets() function unimplemented,
because it will largely depend on the functionality of the binary itself
or the crates used in the project. For a fleshed out example where the
binary can generate its own assets, see `spellout`:

https://github.com/EarthmanMuons/spellout/blob/799e6a42899fcee6d0225588f725649696b4fdfa/xtask/src/dist.rs#L92-L141