Sync documentation with lessons learned from spellout
Since the README files get published to other platforms like crates.io,
we want to use absolute URLs because otherwise the relative links to
things like our LICENSE files will be broken. I also shuffled around
the various sections based on the binary being the main project of the
repository, that relies on a published library.